--- packages/streak/src/components/Button/index.stories.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-
-import { ButtonType, Button } from './index';
-
-export default {
-  title: '버튼',
-  component: Button,
-  parameters: {
-    componentSubtitle: 'Button 컴포넌트',
-  },
-};
-
-export const defaultButton = () => {
-  return <Button>default 버튼</Button>;
-};
-
-export const primaryButton = () => {
-  return <Button theme={ButtonType.PRIMARY}>primary 버튼</Button>;
-};
-export const secondaryButton = () => {
-  return <Button theme={ButtonType.SECONDARY}>secondary 버튼</Button>;
-};
